+++ src/api/daily/sync.daily.kakao.w.ts
@@ -1,20 +1,13 @@
 import axios, {AxiosResponse} from 'axios';
 import {ToonProvider, ToonStatus} from "@prisma/client";
-import {mapGenreToEnum} from "../../genre.to.enum";
-import {ToonDto} from "../../../toon/dto/toon.dto";
-import {SyncResult} from "../../../toon/interface/interface.syncresult";
-import {ToonUpdate} from "../../../toon/interface/interface.toon.update";
-import {findExistsToons} from "../../find.exists.toons";
+import {mapGenreToEnum} from "../../util/genre.to.enum";
+import {ToonDto} from "../../toon/dto/toon.dto";
+import {SyncResult} from "../../toon/interface/interface.syncresult";
+import {ToonUpdate} from "../../toon/interface/interface.toon.update";
+import {findExistsToons} from "../../util/find.exists.toons";
+import {mergeDailyToons} from "../../util/merge.daily.toons";
 
 const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
-
-const COMMON_HEADERS = {
-  'Accept-Language': 'ko',
-  'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/126.0.0.0 Safari/537.36',
-  'Content-Type': 'application/json',
-  'Referer': 'https://page.kakao.com/',
-  'Origin': 'https://page.kakao.com',
-};
 
 const genres =
     ["무협", "로맨스", "스릴러",
@@ -22,9 +15,15 @@ const genres =
       "판타지", "드라마", "스포츠",
       "감성", "로판", "로맨스 판타지"];
 
+const COMMON_HEADERS = {
+  'Accept-Language': 'ko',
+  'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/141.0.0.0 Safari/537.36',
+  'Content-Type': 'application/json',
+};
+
 async function fetchWebtoonData(day: string): Promise<SyncResult> {
   const url = `https://gateway-kw.kakao.com/section/v2/timetables/days?placement=timetable_${day}`;
-  const existingToons = await findExistsToons(ToonProvider.NAVER);
+  const existingToons = await findExistsToons(ToonProvider.KAKAO_W);
 
   let createData: ToonDto[] = [];
   let updateData: ToonUpdate[] = [];
@@ -44,19 +43,20 @@ async function fetchWebtoonData(day: string): Promise<SyncResult> {
 
         if (findToon) {
           if (findToon.status !== ToonStatus.FINISHED || findToon.totalEpisode !== totalEpisode) {
+            console.log(`업데이트 : ${title}`);
             const webtoon = {
               toonId,
               title,
               status: ToonStatus.FINISHED,
               publishDays: '완결',
-              rating: Number(item.starScore.toFixed(1)),
               totalEpisode,
-              provider: ToonProvider.NAVER,
+              provider: ToonProvider.KAKAO_W,
             } as ToonUpdate;
 
             updateData.push(webtoon);
           }
         }else{
+          console.log(`추가 : ${title}`);
           const webtoon = {
             toonId,
             provider: ToonProvider.KAKAO_W,
@@ -75,7 +75,6 @@ async function fetchWebtoonData(day: string): Promise<SyncResult> {
 
           createData.push(webtoon);
         }
-        console.log(`✅ 수집 중: ${item.content.title}`);
 
       } catch (e) {
         console.error(`❌ [ID: ${toonId}] 상세 정보 수집 실패: ${e.message}`);
@@ -143,12 +142,14 @@ async function fetchSummaryData(id: number): Promise<any> {
   }
 }
 
-export async function getKakaoWFinishedToons(): Promise<SyncResult> {
-  console.log("카카오 웹툰 데이터 수집 시작");
+export async function getKakaoWDailyToons(): Promise<SyncResult> {
+  console.log("[데일리] 카카오 웹툰 데이터 수집 시작");
+  const days = ['mon', 'tue', 'wed', 'thu', 'fri', 'sat', 'sun', 'complete'];
+  // const days = ['mon'];
   const startTime = Date.now();
-  const result = await fetchWebtoonData('complete')
+  const result = await mergeDailyToons(days, fetchWebtoonData);
   const endTime = Date.now();
   const duration = ((endTime - startTime) / 1000).toFixed(1);
-  console.log(`카카오 웹툰 수집 완료 - 갯수 ${result.createData.length + result.updateData.length}개, 소요 시간: ${duration}초`);
+  console.log(`[데일리] 카카오 웹툰 수집 완료 - 갯수 ${result.createData.length + result.updateData.length}개, 소요 시간: ${duration}초`);
   return result;
 }
