+++ src/api/finish/sync.finished.naver.ts
@@ -1,10 +1,10 @@
 import axios, {AxiosResponse} from "axios";
 import {ToonProvider, ToonStatus} from "@prisma/client";
-import {mapGenreToEnum} from "../../genre.to.enum";
-import {ToonDto} from "../../../toon/dto/toon.dto";
-import {SyncResult} from "../../../toon/interface/interface.syncresult";
-import {ToonUpdate} from "../../../toon/interface/interface.toon.update";
-import {findExistsToons} from "../../find.exists.toons";
+import {mapGenreToEnum} from "../../util/genre.to.enum";
+import {ToonDto} from "../../toon/dto/toon.dto";
+import {SyncResult} from "../../toon/interface/interface.syncresult";
+import {ToonUpdate} from "../../toon/interface/interface.toon.update";
+import {findExistsToons} from "../../util/find.exists.toons";
 
 const COMMON_HEADERS = {
   'Accept-Language': 'ko',
