+++ src/app.module.ts
@@ -4,7 +4,7 @@ import {AuthModule} from './auth/auth.module';
 import {APP_FILTER, APP_INTERCEPTOR} from "@nestjs/core";
 import {HttpExceptionFilter} from "./exception/http.exception.filter";
 import {LoggerMiddleware} from "./middlewares/logger.middleware";
-import {LoggingInterceptor} from "./success.interceptor";
+import {LoggingInterceptor} from "./interceptors/success.interceptor";
 import {PrismaModule} from "../prisma/prisma.module";
 import {ConfigModule} from "@nestjs/config";
 import {ScheduleModule} from "@nestjs/schedule";
